+++ WebApplicationTravel/Controllers/CountriesController.cs
@@ -15,14 +15,29 @@ namespace WebApplicationTravel.Controllers
         private MSGDBContext db = new MSGDBContext();
 
         // GET: Countries
-        public ActionResult Index()
+        public ActionResult Index(string search)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            var s = from c in db.Countries
+                    select c;
+            if (!String.IsNullOrEmpty(search))
+            {
+                s = s.Where(x => x.CountryName.Equals(search));
+                return View(s.ToList());
+            }
             return View(db.Countries.ToList());
         }
 
         // GET: Countries/Details/5
         public ActionResult Details(int? id)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -48,6 +63,10 @@ namespace WebApplicationTravel.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "CountryId,CountryName")] Country country)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Countries.Add(country);
@@ -61,6 +80,10 @@ namespace WebApplicationTravel.Controllers
         // GET: Countries/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -80,6 +103,10 @@ namespace WebApplicationTravel.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "CountryId,CountryName")] Country country)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(country).State = EntityState.Modified;
@@ -92,6 +119,10 @@ namespace WebApplicationTravel.Controllers
         // GET: Countries/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -109,14 +140,24 @@ namespace WebApplicationTravel.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["Admin"] == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             Country country = db.Countries.Find(id);
+            IEnumerable<City> q = db.Cities.Where(c => c.CountryId==id);
+            foreach (City c in q)
+            {
+                db.Cities.Remove(c);
+                db.SaveChanges();
+            }
             db.Countries.Remove(country);
             db.SaveChanges();
             return RedirectToAction("Index");
         }
 
         protected override void Dispose(bool disposing)
-        {
+        { 
             if (disposing)
             {
                 db.Dispose();
